Merge in MCU16CE/dspic33ck-curiosity-crc-advanced from ~C69672/dspic33ck-curiosity-crc-advanced:feature/CC16SCRIP-8248-2 to develop

* commit '93b1f2301f81475bf0ad383edc11a3a70dd8ce1f': (22 commits)
  Updates to README from PR comments
  Got rid of PTG filepath
  Reverting to previous README format (one Operation section)
  Getting rid of warnings and checked MISRA compliance
  Optimized hardware CRC calculation
  Removing post-processing from benchmarking timing
  Made benchmarking print more clear, reflected change in images
  Trying to make headings clearer
  Additional changes to README
  Suggestions from Gianni to README
  Fixed error in README
  Fixed error in README
  Fix to API documentation
  MISRA compliant
  MISRA compliant macros
  EOD work 1-10
  Updated README order and added Default Operation section
  Removed error in README
  Removed all mentions of data vector and other small fixes
  Attempt at MISRA Compliance
  ...

+++ dspic33ck-curiosity-crc-advanced.X/hardware_crc.h
@@ -59,6 +59,6 @@ void configureHardwareCRC(CRC_SETTINGS settings);
  * @param[in]    inputDataSize - The number of 8-bit words in the input data
  * @return       Returns the calculated CRC value calculated given the input settings
  */
-uint32_t calculateHardwareCRC(CRC_SETTINGS settings, uint8_t inputData[], uint8_t inputDataSize);
+uint32_t calculateHardwareCRC(CRC_SETTINGS settings, uint8_t inputData[], uint32_t inputDataSize);
 
 #endif	/* HARDWARE_CRC_H */
